physical_ai_k3s_proxy_v1/apps/central-hub/requirements.txt
flask
kubernetes
pyyaml
redis
# LLM은 나중에 추가